move partition_set_def logic to job definition
